Share command bar test setup
